requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# AI & ML
openai==1.10.0
faiss-cpu==1.7.4
sentence-transformers==2.3.1
numpy==1.26.3

# Telegram
aiogram==3.3.0

# HTTP клиент
httpx==0.26.0
aiohttp==3.9.1

# База данных
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Логирование
loguru==0.7.2

# Утилиты
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Безопасность
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Валидация
validators==0.22.0

# Markdown
markdown==3.5.2

# Экспорт данных
pandas==2.1.4

# Тестирование
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
